+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/aad/webapi/AADJwtBearerTokenAuthenticationConverter.java
@@ -2,7 +2,8 @@
 // Licensed under the MIT License.
 package com.azure.spring.aad.webapi;
 
-import com.azure.spring.common.AbstractJwtBearerTokenAuthenticationConverter;
+import com.azure.spring.aad.AADOAuth2AuthenticatedPrincipal;
+import com.azure.spring.aad.AbstractJwtBearerTokenAuthenticationConverter;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.OAuth2AuthenticatedPrincipal;
@@ -13,7 +14,8 @@ import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.azure.spring.common.AADJwtGrantedAuthoritiesConverter.DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP;
+import static com.azure.spring.aad.AADJwtGrantedAuthoritiesConverter.DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP;
+import static com.azure.spring.aad.AADJwtGrantedAuthoritiesConverter.DEFAULT_PRINCIPAL_CLAIM_NAME;
 
 /**
  * A {@link Converter} that takes a {@link Jwt} and converts it into a {@link BearerTokenAuthentication}.
@@ -28,9 +30,31 @@ public class AADJwtBearerTokenAuthenticationConverter extends AbstractJwtBearerT
     }
 
     /**
+     * Structure AADJwtBearerTokenAuthenticationConverter with the authority claim.
+     * @param principalClaimName principal claim name
+     * @deprecated Recommended to use others constructor.
+     */
+    @Deprecated
+    public AADJwtBearerTokenAuthenticationConverter(String principalClaimName) {
+        super(principalClaimName, DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP.get(DEFAULT_PRINCIPAL_CLAIM_NAME));
+    }
+
+    /**
+     * Structure AADJwtBearerTokenAuthenticationConverter with the authority claim name and prefix.
+     * @param principalClaimName principal claim name
+     * @param authorityPrefix the prefix name of the authority
+     * @deprecated Recommended to use others constructor.
+     */
+    @Deprecated
+    public AADJwtBearerTokenAuthenticationConverter(String principalClaimName, String authorityPrefix) {
+        super(principalClaimName, authorityPrefix);
+    }
+
+    /**
      * Using spring security provides JwtGrantedAuthoritiesConverter, it can resolve the access token of scp or roles.
      *
      * @param principalClaimName authorities claim name
+     * @param claimToAuthorityPrefixMap the authority name and prefix map
      */
     public AADJwtBearerTokenAuthenticationConverter(String principalClaimName,
                                                     Map<String, String> claimToAuthorityPrefixMap) {