+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/autoconfigure/b2c/AADB2CJwtBearerTokenAuthenticationConverter.java
@@ -2,7 +2,8 @@
 // Licensed under the MIT License.
 package com.azure.spring.autoconfigure.b2c;
 
-import com.azure.spring.common.AbstractJwtBearerTokenAuthenticationConverter;
+import com.azure.spring.aad.AADOAuth2AuthenticatedPrincipal;
+import com.azure.spring.aad.AbstractJwtBearerTokenAuthenticationConverter;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.OAuth2AuthenticatedPrincipal;
@@ -13,7 +14,8 @@ import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.azure.spring.common.AADJwtGrantedAuthoritiesConverter.DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP;
+import static com.azure.spring.aad.AADJwtGrantedAuthoritiesConverter.DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP;
+import static com.azure.spring.aad.AADJwtGrantedAuthoritiesConverter.DEFAULT_PRINCIPAL_CLAIM_NAME;
 
 /**
  * A {@link Converter} that takes a {@link Jwt} and converts it into a {@link BearerTokenAuthentication}.
@@ -25,6 +27,27 @@ public class AADB2CJwtBearerTokenAuthenticationConverter extends AbstractJwtBear
      */
     public AADB2CJwtBearerTokenAuthenticationConverter() {
         this(DEFAULT_PRINCIPAL_CLAIM_NAME, DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP);
+    }
+
+    /**
+     * Structure AADB2CJwtBearerTokenAuthenticationConverter with the authority claim.
+     * @param principalClaimName principal claim name
+     * @deprecated Recommended to use others constructor.
+     */
+    @Deprecated
+    public AADB2CJwtBearerTokenAuthenticationConverter(String principalClaimName) {
+        super(principalClaimName, DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP.get(DEFAULT_PRINCIPAL_CLAIM_NAME));
+    }
+
+    /**
+     * Structure AADB2CJwtBearerTokenAuthenticationConverter with the authority claim name and prefix.
+     * @param principalClaimName principal claim name
+     * @param authorityPrefix the prefix name of the authority
+     * @deprecated Recommended to use others constructor.
+     */
+    @Deprecated
+    public AADB2CJwtBearerTokenAuthenticationConverter(String principalClaimName, String authorityPrefix) {
+        super(principalClaimName, authorityPrefix);
     }
 
     public AADB2CJwtBearerTokenAuthenticationConverter(String principalClaimName,
@@ -40,6 +63,6 @@ public class AADB2CJwtBearerTokenAuthenticationConverter extends AbstractJwtBear
         String name = Optional.ofNullable(principalClaimName)
                               .map(n -> (String) claims.get(n))
                               .orElseGet(() -> (String) claims.get("sub"));
-        return new AADB2COAuth2AuthenticatedPrincipal(headers, claims, authorities, tokenValue, name);
+        return new AADOAuth2AuthenticatedPrincipal(headers, claims, authorities, tokenValue, name);
     }
 }