+++ sdk/spring/azure-spring-boot-starter-active-directory-b2c/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Release History
 
 ## 3.6.0-beta.1 (Unreleased)
-
+### Breaking Changes
+- Deprecate the two construction methods of class AADB2CJwtBearerTokenAuthenticationConverter.
+- Remove class `AADB2COAuth2AuthenticatedPrincipal`, use class `AADOAuth2AuthenticatedPrincipal` instead
 
 ## 3.5.0 (2021-05-24)
 ### New Features