+++ sdk/spring/azure-spring-boot-starter-active-directory/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.6.0-beta.1 (Unreleased)
 ### Breaking Changes
 - Deprecate aad.group.enable-full-list, use aad.group.allowed-group-ids=all instead.
+- Deprecate the two construction methods of class `AADJwtBearerTokenAuthenticationConverter`.
 
 ### New Features
 - Support domain_hint in aad-starter.([#21517](https://github.com/Azure/azure-sdk-for-java/issues/21517))